src/Text2SQL_V2-main/requirements.txt
flask
flask-cors
gunicorn
pandas
matplotlib
langchain
langchain-openai
langchain-google-genai
python-dotenv

